Implement Blinking for both Patrol and Stationary
Used a Animation Event in the Blinking Animations for both enemies using the Animation Window. Added a script called
 Redirector to the Sprite object with the Animator component attached. This would let me get the Carrier Script component
and call a method called  AnimEventAlertBlinkOffc(string message) using the message BlinkAnimationOver. Then
in the Animation FSM window I added an anystate that would us the condition IsBlinked which is set to the Carrier
Script bool HasBlinked. This variable is public for testing purposes. DO NOT CHANGE SPEED OF BLINK OR IT WILL
BREAK. This way, when the Blink is over, it will stop the Blink Animation.