The UI SFX was added
The click sound was created and added to the project's file library. Inside the folder Assets/SFX/UI.
I created null objects named "Button Clicked" within the "background" object of each scene with Audio Source components to load the audio files along with the scene.
Then the "button" command was instructed within each button of the UI, to after a click, instruct the Button Clicked object within the background object of each level to reproduce the button_selected.wav sound.
However, due to Unity's standard UI button's nature, in some cases, the scene is changed before the end of SFX, causing a strangeness. I'm not sure if it is possible to fix this using only the standard Unity button component. It may need to be corrected by converting the commands to a script.